Sync.

+++ trunk/Sloppy/src/Sloppy/Gtk/uihelper.py
@@ -174,7 +174,7 @@ def new_section(frame_title, child):
 
     alignment = gtk.Alignment()
     alignment.set(0.0,0.0,1.0,1.0)
-    alignment.set_padding(0.0,0.0,1.5*SECTION_SPACING,0.0)
+    alignment.set_padding(0,0,int(1.5*SECTION_SPACING),0)
     alignment.add(child)
 
     frame.add(alignment)
